issue #119, issue #111

+++ veejay-current/veejay-server/libel/avilib.h
@@ -333,9 +333,9 @@ long AVI_bytes_remain(avi_t *AVI);
 int  AVI_close(avi_t *AVI);
 long AVI_bytes_written(avi_t *AVI);
 
-avi_t *AVI_open_input_file(char *filename, int getIndex, int mmap_size);
+avi_t *AVI_open_input_file(char *filename, int getIndex, long mmap_size);
 avi_t *AVI_open_input_indexfile(char *filename, int getIndex, char *indexfile);
-avi_t *AVI_open_fd(int fd, int getIndex, int mmap_size);
+avi_t *AVI_open_fd(int fd, int getIndex, long mmap_size);
 avi_t *AVI_open_indexfd(int fd, int getIndex, char *indexfile);
 int avi_parse_input_file(avi_t *AVI, int getIndex);
 int avi_parse_index_from_file(avi_t *AVI, char *filename);
